Read roles from database
Instead of hard-coding all possible values for roles, read names
and ids from existing database. There needs to be a conversion
from Role class to Rol enum, since it's easier to work with a
closed list.

+++ src/data/session_dto.rs
@@ -11,7 +11,7 @@ pub struct CurrentUserDto {
     pub id: i64,
     pub anonymous: bool,
     pub username: String,
-    pub roles: Vec<u8>,
+    pub roles: Vec<i16>,
     pub active: bool,
     pub baby_id: Vec<BabyInfo>,
 }
@@ -21,7 +21,7 @@ impl CurrentUserDto {
         id: i64,
         anonymous: bool,
         username: String,
-        roles: Vec<u8>,
+        roles: Vec<i16>,
         active: bool,
         baby_id: Vec<BabyInfo>,
     ) -> Self {
